Use top-to-bottom macro expansion, a la Clojure + prepare for Hackage publishing
Haskell (usually) doesn't care about the order in which things are defined, but keeping this guarantee in the presence of macros has turned out to be very difficult (if realistically feasible at all). Because macros can be defined by other macros, the AST would need to be re-expanded and re-expanded, and some edge cases were very nontrivial. So, we're following in Clojure's footsteps, which has solved the problem by enforcing top-to-bottom expansion (note that non-macro-involved functions can be defined wherever and are not subject to this restriction).