Rewrite macro expansion, and remove manual copying of AST
All `stack` projects created by `axel` will have `axel` as a dependency
anyway, so we can assume that we have e.g. the `Axel` module available to us for free (but none of, say, `axel`'s dependencies like `lens`). So, we
can now reorganize the code along logical boundaries rather than "put it
all in one big file so that users can have access to all these
definitions."
We've basically always(?) been able to do this, so the hacky "manually
copy in the AST file" strategy has been unnecessary the whole time.
Live and learn 🙃

Oh, and in other news, I rewrote a bunch of the macro expansion code to
make it cleaner and also work correctly(!).
The new macro expansion strategy is as follows:
  1) Expand the program as one big expression (from `begin` down).
  2) Whenever a macro is expanded, restart expansion entirely.
  3) If we manage to get to a top-level statement without having
     restarted along the way, then we're good to go and we add it
     to our environment accumulator.
  4) ... and, that's basically it.
Quoting happens at evaluation-time (well, not quite... it's after macro expansion and before Haskell evaluation), and quasiquotation is expanded at read-time.
There are probably a bunch of edge cases to work out. However, the macro expansion code
is a lot cleaner now, and *much* less Hadoken-shaped (yayayay!!!).

Oh, and scaffold/macro-def-and-env files are now created under a unique ID in an axelTemp folder (to allow for parallel expansion, i.e. when running this project's tests).

So yeah, everything's probably broken and it'll take a while to iron all
the newly-introduced bugs out... but it'll be worth it for the eventual
maintainability (and correctness) benefits.