[WIP] Migrate to Cabal
This is in preparation for the PureScript backend, since the
`purescript` package (i.e. the compiler itself, implemented in Haskell)
is not available on Stackage. Furthermore, adding it to `extra-deps` is
not possible due to sub-dependency version conflicts.

Now that Cabal 3 incorporates many of the advantages of Stack, it isn't
a bad idea to switch back anyway.