Use lists where appropriate in syntactic forms
Wherever lists of things are used in syntactic forms, e.g. bindings in `let <bindings> <body>`, we now wrap them with `[]` instead of `()`. This makes the structure of the command a bit clearer, and also allows for more flexibility (e.g. we could treat a `[]`-delimited list at the end of a `data` form as a list of typeclasses to derive).